Move set_layout and get_layout to global_setting.py
