+++ dynamiqs/qarrays/utils.py
@@ -10,7 +10,7 @@ from .dense_qarray import DenseQArray
 from .sparse_dia_qarray import SparseDIAQArray
 from .types import QArray
 
-__all__ = ['stack', 'set_layout']
+__all__ = ['stack']
 
 
 def stack(qarrays: Sequence[QArray], axis: int = 0) -> QArray:
@@ -82,64 +82,9 @@ def stack(qarrays: Sequence[QArray], axis: int = 0) -> QArray:
             'Stacking qarrays with different types is not implemented.'
         )
 
-
 class Layout(Enum):
     DENSE = 0
     DIA = 1
 
-
 dense = Layout.DENSE
 dia = Layout.DIA
-
-_DEFAULT_LAYOUT = dia
-
-
-def set_layout(layout: Layout):
-    """Configure the default matrix layout for operators supporting this option.
-
-    Two layouts are supported by most operators (see the list of available operators in
-    the [Python API](../../index.md#operators)):
-
-    - `dq.dense`: JAX native dense layout,
-    - `dq.dia`: dynamiqs sparse diagonal layout, only non-zero diagonals are stored.
-
-    Note:
-        The default layout upon importing dynamiqs is `dq.dia`.
-
-    Args:
-        layout: Default matrix layout for operators (`dq.dense` or `dq.dia`).
-
-    Examples:
-        >>> dq.eye(4)
-        SparseDIAQArray: shape=(4, 4), dims=(4,), dtype=complex64, ndiags=1
-        [[1.+0.j   ⋅      ⋅      ⋅   ]
-         [  ⋅    1.+0.j   ⋅      ⋅   ]
-         [  ⋅      ⋅    1.+0.j   ⋅   ]
-         [  ⋅      ⋅      ⋅    1.+0.j]]
-        >>> dq.set_layout(dq.dense)
-        >>> dq.eye(4)
-        DenseQArray: shape=(4, 4), dims=(4,), dtype=complex64
-        [[1.+0.j 0.+0.j 0.+0.j 0.+0.j]
-         [0.+0.j 1.+0.j 0.+0.j 0.+0.j]
-         [0.+0.j 0.+0.j 1.+0.j 0.+0.j]
-         [0.+0.j 0.+0.j 0.+0.j 1.+0.j]]
-    """
-    if not isinstance(layout, Layout):
-        raise TypeError(
-            f'Argument `layout` must be `dq.dense` or `dq.dia`, but is `{layout}`'
-        )
-
-    global _DEFAULT_LAYOUT  # noqa: PLW0603
-    _DEFAULT_LAYOUT = layout
-
-
-def get_layout(layout: Layout | None = None) -> Layout:
-    if layout is None:
-        return _DEFAULT_LAYOUT
-    elif isinstance(layout, Layout):
-        return layout
-    else:
-        raise TypeError(
-            'Argument `layout` must be `dq.dense`, `dq.dia` or `None`, but is'
-            f' `{layout}`.'
-        )
